requirements.txt
discord.py==2.1.0
python-dotenv
